+++ src/front/spv/image.rs
@@ -256,7 +256,7 @@ impl<I: Iterator<Item = u32>> super::Frontend<I> {
         &mut self,
         words_left: u16,
         ctx: &mut super::BlockContext,
-        emitter: &mut crate::front::Emitter,
+        emitter: &mut crate::proc::Emitter,
         block: &mut crate::Block,
         body_idx: usize,
     ) -> Result<crate::Statement, Error> {
@@ -315,7 +315,7 @@ impl<I: Iterator<Item = u32>> super::Frontend<I> {
         &mut self,
         mut words_left: u16,
         ctx: &mut super::BlockContext,
-        emitter: &mut crate::front::Emitter,
+        emitter: &mut crate::proc::Emitter,
         block: &mut crate::Block,
         block_id: spirv::Word,
         body_idx: usize,
@@ -415,7 +415,7 @@ impl<I: Iterator<Item = u32>> super::Frontend<I> {
         mut words_left: u16,
         options: SamplingOptions,
         ctx: &mut super::BlockContext,
-        emitter: &mut crate::front::Emitter,
+        emitter: &mut crate::proc::Emitter,
         block: &mut crate::Block,
         block_id: spirv::Word,
         body_idx: usize,
@@ -663,7 +663,7 @@ impl<I: Iterator<Item = u32>> super::Frontend<I> {
         &mut self,
         at_level: bool,
         ctx: &mut super::BlockContext,
-        emitter: &mut crate::front::Emitter,
+        emitter: &mut crate::proc::Emitter,
         block: &mut crate::Block,
         block_id: spirv::Word,
         body_idx: usize,
